requirements.txt
gradio
requests
PyMuPDF
markdown
python-dotenv
